+++ google-cloud-clients/google-cloud-spanner/src/main/java/com/google/cloud/spanner/SpannerImpl.java
@@ -1420,80 +1420,6 @@ class SpannerImpl extends BaseService<SpannerOptions> implements Spanner {
     }
   }
 
-  static class PartitionedDMLTransaction implements SessionTransaction {
-    private final ByteString transactionId;
-    private final SessionImpl session;
-    private final SpannerRpc rpc;
-    private volatile boolean isValid = true;
-
-    PartitionedDMLTransaction(SessionImpl session, SpannerRpc rpc) {
-      this.session = session;
-      this.rpc = rpc;
-      this.transactionId = initTransaction();
-    }
-
-    ByteString initTransaction() {
-      final BeginTransactionRequest request =
-          BeginTransactionRequest.newBuilder()
-              .setSession(session.getName())
-              .setOptions(
-                  TransactionOptions.newBuilder()
-                      .setPartitionedDml(TransactionOptions.PartitionedDml.getDefaultInstance()))
-              .build();
-      Transaction txn =
-          runWithRetries(
-              new Callable<Transaction>() {
-                @Override
-                public Transaction call() throws Exception {
-                  return rpc.beginTransaction(request, session.options);
-                }
-              });
-      if (txn.getId().isEmpty()) {
-        throw SpannerExceptionFactory.newSpannerException(
-            ErrorCode.INTERNAL,
-            "Failed to init transaction, missing transaction id\n" + session.getName());
-      }
-      return txn.getId();
-    }
-
-    public long executePartitionedUpdate(Statement statement) {
-      checkState(isValid, "Partitioned DML has been invalidated by a new operation on the session");
-      final ExecuteSqlRequest.Builder builder =
-          ExecuteSqlRequest.newBuilder()
-              .setSql(statement.getSql())
-              .setQueryMode(QueryMode.NORMAL)
-              .setSession(session.name)
-              .setTransaction(TransactionSelector.newBuilder().setId(transactionId).build());
-      Map<String, Value> stmtParameters = statement.getParameters();
-      if (!stmtParameters.isEmpty()) {
-        com.google.protobuf.Struct.Builder paramsBuilder = builder.getParamsBuilder();
-        for (Map.Entry<String, Value> param : stmtParameters.entrySet()) {
-          paramsBuilder.putFields(param.getKey(), param.getValue().toProto());
-          builder.putParamTypes(param.getKey(), param.getValue().getType().toProto());
-        }
-      }
-      com.google.spanner.v1.ResultSet resultSet =
-          runWithRetries(
-              new Callable<com.google.spanner.v1.ResultSet>() {
-                @Override
-                public com.google.spanner.v1.ResultSet call() throws Exception {
-                  return rpc.executeQuery(builder.build(), session.options);
-                }
-              });
-      if (!resultSet.hasStats()) {
-        throw new IllegalArgumentException(
-            "Partitioned DML response missing stats possibly due to non-DML statement as input");
-      }
-      // For partitioned DML, using the row count lower bound.
-      return resultSet.getStats().getRowCountLowerBound();
-    }
-
-    @Override
-    public void invalidate() {
-      isValid = false;
-    }
-  }
-
   @VisibleForTesting
   static class TransactionContextImpl extends AbstractReadContext implements TransactionContext {
     @GuardedBy("lock")
